# Microoptimization: Make some functions const