@deeprobin 2021-10-08
> 
> 
> The functions marked as const are used during font loading. The benchmarks you provided show a performance regression after the const changes `load/fontdue`.

Weird. `const` only ensures that the compiler tries to evaluate the method in compile time, otherwise nothing else should happen.

In my opinion, there should be no regression. Possibly something went wrong with my benchmark. Maybe it would be useful if you or someone else provides a benchmark.

### References
* [Documentation: Const Evaluation](https://doc.rust-lang.org/reference/const_eval.html)